Move tabulate to formatters modules
